requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
dnspython==2.1.0
idna==3.1
ivozzo-pyboilerplate==1.0.0
multidict==5.1.0
numpy==1.20.1
pymongo==3.11.3
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
